+++ src/commands/mod.rs
@@ -4,26 +4,39 @@ mod import;
 mod judgement;
 mod past;
 mod ping;
+mod query;
 mod report;
 mod report_message_to_admins;
 mod search;
 
-use std::env::var;
 use anyhow::anyhow;
 use once_cell::sync::Lazy;
+use std::env::var;
 
-use serenity::all::{CommandInteraction, CreateInteractionResponse, CreateInteractionResponseMessage, GuildId, Member, UserId};
+use serenity::all::{
+    CommandInteraction, CreateInteractionResponse, CreateInteractionResponseMessage, GuildId,
+    Member, UserId,
+};
 use serenity::builder::{CreateCommand, CreateInteractionResponseFollowup};
 use serenity::model::{application, Permissions};
 use serenity::prelude::*;
 use tracing::{error, info, instrument};
 
-static DEV_GUILD: Lazy<GuildId> = Lazy::new(|| var("DEV_GUILD").ok().map(|i| GuildId::new(i.parse().expect("Failed to load dev_guild"))).unwrap());
+static DEV_GUILD: Lazy<GuildId> = Lazy::new(|| {
+    var("DEV_GUILD")
+        .ok()
+        .map(|i| GuildId::new(i.parse().expect("Failed to load dev_guild")))
+        .unwrap()
+});
 
 pub async fn run_command(ctx: &Context, interact: &CommandInteraction) {
     let cmd = interact.data.name.as_str();
     let req_perm_level = perm_level(cmd);
-    if !does_fit(interact.member.as_ref().expect("a member"), req_perm_level, ctx) {
+    if !does_fit(
+        interact.member.as_ref().expect("a member"),
+        req_perm_level,
+        ctx,
+    ) {
         let _ = interact
             .create_response(
                 &ctx,
@@ -46,6 +59,7 @@ pub async fn run_command(ctx: &Context, interact: &CommandInteraction) {
         "judgement" => judgement::run(ctx, interact).await,
         "Report Message" => report_message_to_admins::run(ctx, interact).await,
         "search" => search::run(ctx, interact).await,
+        "query" => query::run(ctx, interact).await,
         _ => Err(anyhow!("Unknown command! {}", interact.data.name)),
     } {
         error!("Error running command {}: {:?}", interact.data.name, e);
@@ -58,25 +72,20 @@ pub async fn run_command(ctx: &Context, interact: &CommandInteraction) {
 fn does_fit(user: &Member, perm: PermLevel, ctx: &Context) -> bool {
     match perm {
         PermLevel::Anyone => true,
-        PermLevel::Staff => {
-            user.roles
-                .iter()
-                .any(|role| {
-                    role.to_role_cached(ctx).is_some_and(|e| {
-                        e.permissions.contains(Permissions::ADMINISTRATOR)
-                            || e.name.contains("Admin")
-                            || e.name.contains("Mod")
-                    })
-                })
-        },
+        PermLevel::Staff => user.roles.iter().any(|role| {
+            role.to_role_cached(ctx).is_some_and(|e| {
+                e.permissions.contains(Permissions::ADMINISTRATOR)
+                    || e.name.contains("Admin")
+                    || e.name.contains("Mod")
+            })
+        }),
         PermLevel::Dev => user.user.id == OWNER,
-        PermLevel::God | PermLevel::Nobody | PermLevel::Invalid => false
+        PermLevel::God | PermLevel::Nobody | PermLevel::Invalid => false,
     }
 }
 
-
 const OWNER: UserId = UserId::new(171629704959229952);
-use strum::{EnumVariantNames, EnumString, AsRefStr};
+use strum::{AsRefStr, EnumString, EnumVariantNames};
 #[derive(Debug, PartialEq, Eq, EnumVariantNames, EnumString, AsRefStr)]
 enum PermLevel {
     Anyone,
@@ -84,17 +93,18 @@ enum PermLevel {
     Dev,
     God,
     Nobody,
-    Invalid
+    Invalid,
 }
 
 fn perm_level(command: &str) -> PermLevel {
     use crate::commands::PermLevel::*;
     match command {
         "ping" | "Report Message" => Anyone,
-        "audit" | "discord" | "Audit Message" | "Audit User" | "past" | "report" | "search" => Staff,
+        "audit" | "discord" | "Audit Message" | "Audit User" | "past" | "report" | "search"
+        | "query" => Staff,
         "import" => Dev,
         "when" => God,
-        _ => Invalid
+        _ => Invalid,
     }
 }
 
@@ -102,14 +112,12 @@ fn do_perms(c: CreateCommand) -> CreateCommand {
     c.dm_permission(false)
 }
 
-
 macro_rules! register_command {
     ($ctx:expr, $cmd:expr) => {
         if cfg!(debug_assertions) {
             DEV_GUILD.create_command($ctx, $cmd).await
         } else {
-            application::Command::create_global_command($ctx, $cmd)
-            .await
+            application::Command::create_global_command($ctx, $cmd).await
         }
     };
 }
@@ -118,12 +126,16 @@ macro_rules! command {
     ($ctx:expr, $name:ident) => {
         let (command, name) = $name::register();
         let perms = perm_level(name);
-        info!("Registering command: {} with perms: {:?}", stringify!($name), perms);
+        info!(
+            "Registering command: {} with perms: {:?}",
+            stringify!($name),
+            perms
+        );
         assert_ne!(perms, crate::commands::PermLevel::Invalid);
         register_command!($ctx, do_perms(command)).expect(&format!(
-                "Failed to create global command {}",
-                stringify!($name)
-            ));
+            "Failed to create global command {}",
+            stringify!($name)
+        ));
     };
     ($ctx:expr, $name:ident, $sub:ident) => {
         let (command, name) = $name::$sub::register();
@@ -135,12 +147,11 @@ macro_rules! command {
             perms
         );
         assert_ne!(perms, crate::commands::PermLevel::Invalid);
-        register_command!($ctx, do_perms(command))
-            .expect(&format!(
-                "Failed to create global command {}::{}",
-                stringify!($name),
-                stringify!($sub)
-            ))
+        register_command!($ctx, do_perms(command)).expect(&format!(
+            "Failed to create global command {}::{}",
+            stringify!($name),
+            stringify!($sub)
+        ))
     };
 }
 #[instrument(skip(ctx))]
@@ -160,13 +171,17 @@ pub async fn register_commands(ctx: &impl CacheHttp) {
     for command in commands {
         info!("Unregistering command {}", command.name);
         if cfg!(debug_assertions) {
-            DEV_GUILD.delete_command(ctx.http(), command.id).await.unwrap();
+            DEV_GUILD
+                .delete_command(ctx.http(), command.id)
+                .await
+                .unwrap();
         } else {
             application::Command::delete_global_command(ctx.http(), command.id)
                 .await
                 .unwrap();
         }
     }
+
     command!(ctx, ping);
     command!(ctx, audit);
     command!(ctx, audit_discord, user);
@@ -178,5 +193,6 @@ pub async fn register_commands(ctx: &impl CacheHttp) {
     command!(ctx, judgement);
     command!(ctx, report_message_to_admins);
     command!(ctx, search);
+    command!(ctx, query);
     info!("All commands registered!")
 }