+++ src/commands/report.rs
@@ -45,10 +45,17 @@ pub async fn run(ctx: &Context, interaction: &CommandInteraction) -> anyhow::Res
 }
 
 pub fn register() -> (CreateCommand, &'static str) {
-    (CreateCommand::new("report")
-        .description("Get a report!")
-        .add_option(
-            CreateCommandOption::new(CommandOptionType::Integer, "report_id", "Id of the report")
+    (
+        CreateCommand::new("report")
+            .description("Get a report!")
+            .add_option(
+                CreateCommandOption::new(
+                    CommandOptionType::Integer,
+                    "report_id",
+                    "Id of the report",
+                )
                 .required(true),
-        ), "report")
+            ),
+        "report",
+    )
 }